# Adding an optimization for dots.