This is nearly trivial but it helps in the common case where there is '.' in the path but there is neither a single-dot nor a double-dot path segment.

On x64 with GCC 11:


This PR:

```
BasicBench_AdaURL_With_Move       3788 ns         3787 ns       188401 GHz=3.45727 cycle/byte=15.6118 cycles/url=1.497k instructions/byte=32.1333 instructions/cycle=2.05826 instructions/ns=7.11599 instructions/url=3.08122k ns/url=433 speed=227.868M/s time/byte=4.38851ns time/url=420.809ns url/s=2.37637M/s
```

Main:

```
BasicBench_AdaURL_With_Move       5196 ns         5196 ns       132032 GHz=3.43774 cycle/byte=21.1124 cycles/url=2.02444k instructions/byte=42.9038 instructions/cycle=2.03216 instructions/ns=6.98604 instructions/url=4.114k ns/url=588.889 speed=166.103M/s time/byte=6.02035ns time/url=577.285ns url/s=1.73225M/s
```